finished m2

+++ src/components/Note.js
@@ -2,8 +2,11 @@ import React from "react";
 
 const Note = ({ note, deleteNote }) => {
     console.log(note.text);
-    const handleDeleteNote = () =>{
-        deleteNote(note.id)
+    function handleDeleteNote(){
+        const result = confirm("Are you sure you want to delete?")
+        if(result){
+         deleteNote(note.id)
+        }
 
     }
     return(
@@ -11,7 +14,7 @@ const Note = ({ note, deleteNote }) => {
         <div className="noteInfo">
             <div className="noteData">{note.text}</div>
             <h3 className="noteData">{note.date.toString()}</h3>
-            <button onClick = {handleDeleteNote}>Delete Note</button>
+            <button className="deleteButton" onClick = {handleDeleteNote}>Delete Note</button>
         </div>
         </div>
     )
